backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
ifcopenshell==0.8.3.post2
pydantic==2.3.0
confluent-kafka==2.3.0
pymongo==4.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.0
requests==2.31.0 